rename files to readers
