+++ web/clients/grpc-web/make_service_client.ts
@@ -1,4 +1,4 @@
-import { MakeReaderClient as MakeReaderClientGrpc } from "../../protobuf/make/Make_serviceServiceClientPb";
+import { MakeReaderClient as MakeReaderClientGrpc } from "../../protobuf/make/Make_readerServiceClientPb";
 
 const MakeReaderClient = new MakeReaderClientGrpc("http://localhost:8080");
