+++ web/clients/grpc-web/phone_service_client.ts
@@ -1,4 +1,4 @@
-import { PhoneReaderClient as PhoneReaderClientGrpc } from "../../protobuf/phone/Phone_serviceServiceClientPb";
+import { PhoneReaderClient as PhoneReaderClientGrpc } from "../../protobuf/phone/Phone_readerServiceClientPb";
 
 const PhoneReaderClient = new PhoneReaderClientGrpc("http://localhost:8080");
