+++ web/pages/make/index.tsx
@@ -6,7 +6,7 @@ import Container from "../../components/container";
 import SEO from "../../components/seo";
 import MakeReaderClient from "../../clients/grpc-web/make_service_client";
 import { Make } from "../../protobuf/make/make_pb";
-import { ListByPageRequest } from "../../protobuf/make/make_service_pb";
+import { ListByPageRequest } from "../../protobuf/make/make_reader_pb";
 
 import { ListByPageClientSide, PageResult } from "../../components/listPage";
