+++ web/pages/phone/[id].tsx
@@ -3,7 +3,7 @@ import { StatusCode } from "grpc-web";
 import Container from "../../components/container";
 import SEO from "../../components/seo";
 import Custom404 from "../404";
-import { GetOneByIDRequest } from "../../protobuf/phone/phone_service_pb";
+import { GetOneByIDRequest } from "../../protobuf/phone/phone_reader_pb";
 import { Phone } from "../../protobuf/phone/phone_pb";
 import PhoneReaderClient from "../../clients/nodejs/phone_service_client";
 import PhoneComponent from "../../components/phone";
