+++ web/protobuf/os/Os_readerServiceClientPb.ts
@@ -13,7 +13,7 @@
 
 import * as grpcWeb from 'grpc-web';
 
-import * as protobuf_os_os_service_pb from '../../protobuf/os/os_service_pb';
+import * as protobuf_os_os_reader_pb from '../../protobuf/os/os_reader_pb';
 
 
 export class OSReaderClient {
@@ -36,28 +36,28 @@ export class OSReaderClient {
   }
 
   methodInfoGetOneByID = new grpcWeb.AbstractClientBase.MethodInfo(
-    protobuf_os_os_service_pb.GetOneByIDResponse,
-    (request: protobuf_os_os_service_pb.GetOneByIDRequest) => {
+    protobuf_os_os_reader_pb.GetOneByIDResponse,
+    (request: protobuf_os_os_reader_pb.GetOneByIDRequest) => {
       return request.serializeBinary();
     },
-    protobuf_os_os_service_pb.GetOneByIDResponse.deserializeBinary
+    protobuf_os_os_reader_pb.GetOneByIDResponse.deserializeBinary
   );
 
   getOneByID(
-    request: protobuf_os_os_service_pb.GetOneByIDRequest,
-    metadata: grpcWeb.Metadata | null): Promise<protobuf_os_os_service_pb.GetOneByIDResponse>;
+    request: protobuf_os_os_reader_pb.GetOneByIDRequest,
+    metadata: grpcWeb.Metadata | null): Promise<protobuf_os_os_reader_pb.GetOneByIDResponse>;
 
   getOneByID(
-    request: protobuf_os_os_service_pb.GetOneByIDRequest,
+    request: protobuf_os_os_reader_pb.GetOneByIDRequest,
     metadata: grpcWeb.Metadata | null,
     callback: (err: grpcWeb.Error,
-               response: protobuf_os_os_service_pb.GetOneByIDResponse) => void): grpcWeb.ClientReadableStream<protobuf_os_os_service_pb.GetOneByIDResponse>;
+               response: protobuf_os_os_reader_pb.GetOneByIDResponse) => void): grpcWeb.ClientReadableStream<protobuf_os_os_reader_pb.GetOneByIDResponse>;
 
   getOneByID(
-    request: protobuf_os_os_service_pb.GetOneByIDRequest,
+    request: protobuf_os_os_reader_pb.GetOneByIDRequest,
     metadata: grpcWeb.Metadata | null,
     callback?: (err: grpcWeb.Error,
-               response: protobuf_os_os_service_pb.GetOneByIDResponse) => void) {
+               response: protobuf_os_os_reader_pb.GetOneByIDResponse) => void) {
     if (callback !== undefined) {
       return this.client_.rpcCall(
         this.hostname_ +
@@ -76,28 +76,28 @@ export class OSReaderClient {
   }
 
   methodInfoGetManyByIDs = new grpcWeb.AbstractClientBase.MethodInfo(
-    protobuf_os_os_service_pb.GetManyByIDsResponse,
-    (request: protobuf_os_os_service_pb.GetManyByIDsRequest) => {
+    protobuf_os_os_reader_pb.GetManyByIDsResponse,
+    (request: protobuf_os_os_reader_pb.GetManyByIDsRequest) => {
       return request.serializeBinary();
     },
-    protobuf_os_os_service_pb.GetManyByIDsResponse.deserializeBinary
+    protobuf_os_os_reader_pb.GetManyByIDsResponse.deserializeBinary
   );
 
   getManyByIDs(
-    request: protobuf_os_os_service_pb.GetManyByIDsRequest,
-    metadata: grpcWeb.Metadata | null): Promise<protobuf_os_os_service_pb.GetManyByIDsResponse>;
+    request: protobuf_os_os_reader_pb.GetManyByIDsRequest,
+    metadata: grpcWeb.Metadata | null): Promise<protobuf_os_os_reader_pb.GetManyByIDsResponse>;
 
   getManyByIDs(
-    request: protobuf_os_os_service_pb.GetManyByIDsRequest,
+    request: protobuf_os_os_reader_pb.GetManyByIDsRequest,
     metadata: grpcWeb.Metadata | null,
     callback: (err: grpcWeb.Error,
-               response: protobuf_os_os_service_pb.GetManyByIDsResponse) => void): grpcWeb.ClientReadableStream<protobuf_os_os_service_pb.GetManyByIDsResponse>;
+               response: protobuf_os_os_reader_pb.GetManyByIDsResponse) => void): grpcWeb.ClientReadableStream<protobuf_os_os_reader_pb.GetManyByIDsResponse>;
 
   getManyByIDs(
-    request: protobuf_os_os_service_pb.GetManyByIDsRequest,
+    request: protobuf_os_os_reader_pb.GetManyByIDsRequest,
     metadata: grpcWeb.Metadata | null,
     callback?: (err: grpcWeb.Error,
-               response: protobuf_os_os_service_pb.GetManyByIDsResponse) => void) {
+               response: protobuf_os_os_reader_pb.GetManyByIDsResponse) => void) {
     if (callback !== undefined) {
       return this.client_.rpcCall(
         this.hostname_ +
@@ -116,28 +116,28 @@ export class OSReaderClient {
   }
 
   methodInfoListByCursor = new grpcWeb.AbstractClientBase.MethodInfo(
-    protobuf_os_os_service_pb.ListByCursorResponse,
-    (request: protobuf_os_os_service_pb.ListByCursorRequest) => {
+    protobuf_os_os_reader_pb.ListByCursorResponse,
+    (request: protobuf_os_os_reader_pb.ListByCursorRequest) => {
       return request.serializeBinary();
     },
-    protobuf_os_os_service_pb.ListByCursorResponse.deserializeBinary
+    protobuf_os_os_reader_pb.ListByCursorResponse.deserializeBinary
   );
 
   listByCursor(
-    request: protobuf_os_os_service_pb.ListByCursorRequest,
-    metadata: grpcWeb.Metadata | null): Promise<protobuf_os_os_service_pb.ListByCursorResponse>;
+    request: protobuf_os_os_reader_pb.ListByCursorRequest,
+    metadata: grpcWeb.Metadata | null): Promise<protobuf_os_os_reader_pb.ListByCursorResponse>;
 
   listByCursor(
-    request: protobuf_os_os_service_pb.ListByCursorRequest,
+    request: protobuf_os_os_reader_pb.ListByCursorRequest,
     metadata: grpcWeb.Metadata | null,
     callback: (err: grpcWeb.Error,
-               response: protobuf_os_os_service_pb.ListByCursorResponse) => void): grpcWeb.ClientReadableStream<protobuf_os_os_service_pb.ListByCursorResponse>;
+               response: protobuf_os_os_reader_pb.ListByCursorResponse) => void): grpcWeb.ClientReadableStream<protobuf_os_os_reader_pb.ListByCursorResponse>;
 
   listByCursor(
-    request: protobuf_os_os_service_pb.ListByCursorRequest,
+    request: protobuf_os_os_reader_pb.ListByCursorRequest,
     metadata: grpcWeb.Metadata | null,
     callback?: (err: grpcWeb.Error,
-               response: protobuf_os_os_service_pb.ListByCursorResponse) => void) {
+               response: protobuf_os_os_reader_pb.ListByCursorResponse) => void) {
     if (callback !== undefined) {
       return this.client_.rpcCall(
         this.hostname_ +
@@ -156,28 +156,28 @@ export class OSReaderClient {
   }
 
   methodInfoListByPage = new grpcWeb.AbstractClientBase.MethodInfo(
-    protobuf_os_os_service_pb.ListByPageResponse,
-    (request: protobuf_os_os_service_pb.ListByPageRequest) => {
+    protobuf_os_os_reader_pb.ListByPageResponse,
+    (request: protobuf_os_os_reader_pb.ListByPageRequest) => {
       return request.serializeBinary();
     },
-    protobuf_os_os_service_pb.ListByPageResponse.deserializeBinary
+    protobuf_os_os_reader_pb.ListByPageResponse.deserializeBinary
   );
 
   listByPage(
-    request: protobuf_os_os_service_pb.ListByPageRequest,
-    metadata: grpcWeb.Metadata | null): Promise<protobuf_os_os_service_pb.ListByPageResponse>;
+    request: protobuf_os_os_reader_pb.ListByPageRequest,
+    metadata: grpcWeb.Metadata | null): Promise<protobuf_os_os_reader_pb.ListByPageResponse>;
 
   listByPage(
-    request: protobuf_os_os_service_pb.ListByPageRequest,
+    request: protobuf_os_os_reader_pb.ListByPageRequest,
     metadata: grpcWeb.Metadata | null,
     callback: (err: grpcWeb.Error,
-               response: protobuf_os_os_service_pb.ListByPageResponse) => void): grpcWeb.ClientReadableStream<protobuf_os_os_service_pb.ListByPageResponse>;
+               response: protobuf_os_os_reader_pb.ListByPageResponse) => void): grpcWeb.ClientReadableStream<protobuf_os_os_reader_pb.ListByPageResponse>;
 
   listByPage(
-    request: protobuf_os_os_service_pb.ListByPageRequest,
+    request: protobuf_os_os_reader_pb.ListByPageRequest,
     metadata: grpcWeb.Metadata | null,
     callback?: (err: grpcWeb.Error,
-               response: protobuf_os_os_service_pb.ListByPageResponse) => void) {
+               response: protobuf_os_os_reader_pb.ListByPageResponse) => void) {
     if (callback !== undefined) {
       return this.client_.rpcCall(
         this.hostname_ +
