+++ web/protobuf/os/os_reader_pb.d.ts
